refactor(installer): update folder structure
