+++ src/renderer/src/features/installer/components/feature-carousel.tsx
