+++ src/renderer/src/features/installer/components/installation-progress.tsx
