+++ src/renderer/src/features/installer/components/installer-content.stories.tsx
