+++ src/renderer/src/features/installer/components/installer-content.tsx
