+++ src/renderer/src/features/installer/components/installer-styled.tsx
