+++ src/renderer/src/features/installer/components/logo-section.tsx
