+++ src/renderer/src/features/installer/index.ts
@@ -1,5 +1,5 @@
-export * from "./installer-styled";
-export * from "./installer-content";
-export * from "./feature-carousel";
-export * from "./installation-progress";
-export * from "./logo-section";
+export * from "./components/installer-styled";
+export * from "./components/installer-content";
+export * from "./components/feature-carousel";
+export * from "./components/installation-progress";
+export * from "./components/logo-section";
