+++ src/renderer/src/installer-app.tsx
@@ -1,8 +1,8 @@
 import { CssBaseline } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import type React from "react";
-import { InstallerContent } from "./features/installer/installer-content";
-import { AppContainer } from "./features/installer/installer-styled";
+import { InstallerContent } from "./features/installer/components/installer-content";
+import { AppContainer } from "./features/installer/components/installer-styled";
 import theme from "./theme";
 
 /**
